add otterkeys tool
Nyan nyan !

I'm wondering if it would be possible to add my "OtterKeys" to your list of dotnet tools.

OtterKeys lets you quickly generate a new Ed25519 cryptographic key pair, or recover a lost public key. Ed25519 are used to sign and verify a message or other piece of data.